tough: simplify repository load
Introduces a builder object, RepositoryLoader, that takes root,
metadata_base_url, and targets_base_url in its new function. After that
the user can add optional settings such as transport, limits, etc.
Finally, the user calls the load function to load a repository.

The load function on the Repository struct is now private so that the
user does not have to navigate multiple methods of loading a repository.