tough-kms: pad short signatures
This is a critical fix that prevents tough from creating repos that it
cannot load when using tough-kms to sign the repo.

Sometimes KMS returns a signature that is shorter than the key modulus.
Both KMS and OpenSSL will validate such signatures, but ring requires
that the signature length exactly match the modulus. If we pad such
signatures by adding zeros to the beginning until the signature length
matches the modulus length, then we can ensure ring and any other
implementations with the length requirement will be able to validate
the signature.